requirements.txt
asgiref==3.2.7
Django==3.0.5
django-crispy-forms==1.9.0
pdfkit==0.6.1
Pillow==7.1.2
pytz==2019.3
sqlparse==0.3.1
